pyproject.toml
[project]
name = "grafap"
version = "1.0"
description = "Python package that acts as a wrapper for the Microsoft Graph API."
readme = "README.md"
authors = [{ name = "Jordan Maynor", email = "jmaynor@pepsimidamerica.com" }]
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "pyjwt>=2.10.1",
    "pyopenssl>=25.1.0",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: The Unlicense (Unlicense)",
    "Operating System :: OS Independent",
]
license = { file = "LICENSE" }

[project.urls]
Homepage = "https://github.com/pepsimidamerica/grafap"

[build-system]
requires = ["uv_build>=0.7.19,<0.10"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "grafap"
module-root = ""

[dependency-groups]
dev = ["python-dotenv>=1.1.0"]
